+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Added color highlight for problematic properties
 - Upgraded codemirror from version 5 to version 6
 - Add code editor to allow editing the entire style
+- Add support for sprite object in setting modal
 - _...Add new stuff here..._
 
 ### 🐞 Bug fixes